new structure for creating word documents
